backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
PyMuPDF>=1.24.0
python-docx>=1.1.2
requests>=2.31.0
numpy>=1.26.0
python-dotenv>=1.0.1
scikit-learn>=1.4
scipy>=1.11

